refactor: cypress geolocation mock into function
