+++ src/server/lib/getGeolocationRegion.ts
@@ -2,27 +2,16 @@ import { GeoLocation } from '@/shared/model/Geolocation';
 import { Request } from 'express';
 import { Europe } from '@/server/lib/getRegistrationLocation';
 import { CountryCode } from '@guardian/libs';
+import { maybeGetCountryCodeFromCypressMockStateCookie } from '@/server/lib/cypress';
 
 export const getGeolocationRegion = (req: Request): GeoLocation => {
 	/**
 	 * Cypress Test START
-	 *
-	 * This code checks if we're running in Cypress
 	 */
-	const runningInCypress = process.env.RUNNING_IN_CYPRESS === 'true';
-	if (runningInCypress) {
-		const cypressMockStateCookie = req.cookies['cypress-mock-state'];
-
-		// check if the cookie value is an expected value
-		const validCode = ['FR', 'GB', 'US', 'AU'].includes(cypressMockStateCookie);
-
-		// if it is then return the code
-		if (validCode) {
-			return countryCodeToRegion(cypressMockStateCookie);
-		}
-
-		// otherwise let it fall through to the default check below
-	}
+	const maybeMockedCountryCode =
+		maybeGetCountryCodeFromCypressMockStateCookie(req);
+	if (maybeMockedCountryCode)
+		return countryCodeToRegion(maybeMockedCountryCode);
 	/**
 	 * Cypress Test END
 	 */
@@ -30,17 +19,15 @@ export const getGeolocationRegion = (req: Request): GeoLocation => {
 	return countryCodeToRegion(header);
 };
 
-const countryCodeToRegion = (
+export const countryCodeToRegion = (
 	countryCode: string | string[] | undefined,
 ): GeoLocation => {
 	if (Europe.includes(countryCode as CountryCode)) return 'EU';
 	switch (countryCode) {
 		case 'GB':
-			return 'GB';
 		case 'US':
-			return 'US';
 		case 'AU':
-			return 'AU';
+			return countryCode as GeoLocation;
 		default:
 			return 'ROW';
 	}
